or - Copy/services/python-rag/requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.37.1
pydantic==2.8.2 